chore: Update vetrina store and components
The vetrina store is updated to include the setVetrina and clearVetrina actions. The LogoSito.vue component is updated to add an onClick event for setting the vetrina. The AreaRicerca.vue component is updated to import the vetrinaInstance from the vetrinaStore and set the vetrina data on mounted.
Created new folder to keep the homepage and storefront components